Make sure editable requirements install correctly
MAke sure when we specify an editable requirement in our requirements it's installed correctly.
Previously it would just install the requirement as a regular package, meaning certain files would not be
placed (in our case, some fixture of the project we specified). Instead, make sure that if the `-e` flag
is specified the requirement is installed as an editable requirement. Which will include all files of the
requirement.

Regular installing of requirements is still possible, in those cases don't prepend with the `-e` flag.