Allow pip environment markers
Handle pip environment markers correctly. These allow you to specify a
certain requirement for a certain environment and looks like:
```
my-requirement==0.1.2 ; python_version=="3.7"
```
or
```
-e 'git+git@github.com:ByteInternet/my-repo.git@20201127.1#egg=my-repo ; python_version=="3.7"'
```
Notice that we can both handle regular and editable/custom
requirements. Tested with an example requirements file that had several
combinations of this, and it works (only installs that one specified in
the requirements based on the environment).