Do not break on inline comments with environment markers
This already worked for the editable requirements, but not for regular
requirements with environment markers (we just pasted all the token
together, but if there was a comment in there we would accidentally act
like that was part of the requirement). So just make sure we use the
same function in all cases which neatly appends the environment marker.

Also added some tests for this for safety's sake.